Fix: 좌석, 콘서트 도메인 서비스로 이동 시 추가 작업
