+++ src/main/java/kr/hhplus/be/server/domain/concert/ConcertService.java
@@ -3,8 +3,11 @@ package kr.hhplus.be.server.domain.concert;
 import kr.hhplus.be.server.common.exception.AvailableSeatNotFoundException;
 import kr.hhplus.be.server.common.exception.ConcertScheduleNotFoundException;
 import kr.hhplus.be.server.common.exception.ScheduleNotFoundException;
+import kr.hhplus.be.server.common.exception.SeatNotFoundException;
+import kr.hhplus.be.server.domain.dto.SeatScheduleResult;
 import kr.hhplus.be.server.domain.seat.Seat;
 import kr.hhplus.be.server.domain.seat.SeatRepository;
+import kr.hhplus.be.server.domain.seat.SeatStatus;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -35,6 +38,25 @@ public class ConcertService {
         return schedule;
     }
 
+    public Seat updateSeatStatus(Long seatId, SeatStatus seatStatus) {
+        //1. 좌석 정보 조회
+        Seat seat = seatRepository.findByIdWithLock(seatId)
+                .orElseThrow(()-> new SeatNotFoundException("좌석을 찾을 수 없습니다."));
+        //2. 좌석 정보 업데이트
+        Seat updatedSeat = seat.update(seat, seatStatus);
+        //3. 좌석 정보 저장
+        return seatRepository.save(updatedSeat);
+    }
+
+    public Schedule updateScheduleRemainingTicket(Long scheduleId, int increaseOrDecreaseNumber) {
+        Schedule schedule = scheduleRepository.findById(scheduleId)
+                .orElseThrow(()-> new ScheduleNotFoundException("스케줄 정보를 찾을 수 없습니다."));
+
+        Schedule updatedSchedule = schedule.update(schedule, increaseOrDecreaseNumber);
+        return scheduleRepository.save(updatedSchedule);
+    }
+
+
     @Transactional
     public Concert saveConcertWithSchedules(String concertName, List<Schedule> schedules) {
         // Concert 객체 생성
@@ -51,15 +73,6 @@ public class ConcertService {
 
         // Concert와 연관된 Schedule 모두 저장
         return concertRepository.save(concert);
-    }
-
-    @Transactional
-    public Schedule updateScheduleRemainingTicket(Long scheduleId, int increaseOrDecreaseNumber) {
-        Schedule schedule = scheduleRepository.findById(scheduleId)
-                .orElseThrow(()-> new ScheduleNotFoundException("스케줄 정보를 찾을 수 없습니다."));
-
-        Schedule updatedSchedule = schedule.update(schedule, increaseOrDecreaseNumber);
-        return scheduleRepository.save(updatedSchedule);
     }
 
     public Schedule findById(Long scheduleId) {
